vc4graphics.cfg: build DRM_VC4 as a module
Somewhere inside the raspberry code this is expected to be built as a module.
Since i don't know better i assume that is for a reason. But if not at least this misleading error gets removed from the log:

You have a RaspberryPi, but the vc4 Module is not loaded! Set 'dtoverlay=vc4-kms-v3d' in /boot/config.txt and reboot.